It's a generator of Relisting CSVs
Pretty simple little program. 209 lines of code, a good portion of that
is comments. If things don't make sense, gimme a shout. You click the
button and it makes you a CSV with its generated name based on the date,
and puts it in the folder shown in the text box. Click the text box to
open a folder selector, and when one is picked, it saves it to the
application settings so you won't have to again. Any issues in it, it
tells you through the label below the button - bad file names,
nonexistent folders, preexisting files, etc. If all's good, you click
the button and it takes 2-3 mins to run a query before automatically
saving and opening the CSV.