"" swapped for '' and another change
So it didn't like "" for the Check, but '' is fine. Makes sense to be
fair, should have 'checked' on that myself! Also, I have chenged
itemtitle for labelshort as we really should be getting used to seeing
the same title as is on the item, not on the listing. It is for internal
use, not external after all.